ref: made ModalityCapability an unified registry
