+++ fedot_ind/core/models/nn/enums.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class EncoderFamily(str, Enum):
+    """Supported encoder families."""
+
+    cnn = "cnn"
+    mlp = "mlp"
